--- a/libraries/bxdf/gltf_pbr.mtlx
+++ b/libraries/bxdf/gltf_pbr.mtlx
@@ -13,6 +13,9 @@
     <input name="alpha" type="float" value="1" uimin="0" uimax="1" uiname="Alpha" uifolder="Alpha" />
     <input name="alpha_mode" uniform="true" type="integer" enum="OPAQUE, MASK, BLEND" enumvalues="0, 1, 2" value="0" uiname="Alpha Mode" uifolder="Alpha" />
     <input name="alpha_cutoff" uniform="true" type="float" value="0.5" uimin="0" uimax="1" uiname="Alpha Cutoff" uifolder="Alpha" />
+    <input name="iridescence" type="float" value="0" uimin="0" uimax="1" uiname="Iridescence" uifolder="Iridescence" />
+    <input name="iridescence_ior" uniform="true" type="float" value="1.3" uimin="1" uisoftmax="3" uiname="Iridescence Index of Refraction" uifolder="Iridescence" />
+    <input name="iridescence_thickness" type="float" value="100" uimin="0" uisoftmin="100" uisoftmax="400" uiname="Iridescence Thickness" uifolder="Iridescence" />
     <input name="sheen_color" type="color3" value="0, 0, 0" uimin="0, 0, 0" uimax="1, 1, 1" uiname="Sheen Color" uifolder="Sheen" />
     <input name="sheen_roughness" type="float" value="0" uimin="0" uimax="1" uiname="Sheen Roughness" uifolder="Sheen" />
     <input name="clearcoat" type="float" value="0" uimin="0" uimax="1" uiname="Clearcoat" uifolder="Clearcoat" />
@@ -147,11 +150,45 @@
       <input name="normal" type="vector3" interfacename="normal" />
     </generalized_schlick_bsdf>
 
+    <!-- Thin Film -->
+
+    <thin_film_bsdf name="dieletric_thinfilm_bsdf" type="BSDF">
+      <input name="thickness" type="float" interfacename="iridescence_thickness" />
+      <input name="ior" type="float" interfacename="iridescence_ior" />
+    </thin_film_bsdf>
+
+    <layer name="iridescent_dielectric_bsdf" type="BSDF">
+      <input name="top" type="BSDF" nodename="dieletric_thinfilm_bsdf" />
+      <input name="base" type="BSDF" nodename="dielectric_bsdf" />
+    </layer>
+
+    <mix name="mix_iridescent_dielectric_bsdf" type="BSDF">
+      <input name="bg" type="BSDF" nodename="dielectric_bsdf" />
+      <input name="fg" type="BSDF" nodename="iridescent_dielectric_bsdf" />
+      <input name="mix" type="float" interfacename="iridescence" />
+    </mix>
+
+    <thin_film_bsdf name="metal_thinfilm_bsdf" type="BSDF">
+      <input name="thickness" type="float" interfacename="iridescence_thickness" />
+      <input name="ior" type="float" interfacename="iridescence_ior" />
+    </thin_film_bsdf>
+
+    <layer name="iridescent_metal_bsdf" type="BSDF">
+      <input name="top" type="BSDF" nodename="metal_thinfilm_bsdf" />
+      <input name="base" type="BSDF" nodename="metal_bsdf" />
+    </layer>
+
+    <mix name="mix_iridescent_metal_bsdf" type="BSDF">
+      <input name="bg" type="BSDF" nodename="metal_bsdf" />
+      <input name="fg" type="BSDF" nodename="iridescent_metal_bsdf" />
+      <input name="mix" type="float" interfacename="iridescence" />
+    </mix>
+
     <!-- Dielectric/metal mix -->
 
     <mix name="base_mix" type="BSDF">
-      <input name="bg" type="BSDF" nodename="dielectric_bsdf" />
-      <input name="fg" type="BSDF" nodename="metal_bsdf" />
+      <input name="bg" type="BSDF" nodename="mix_iridescent_dielectric_bsdf" />
+      <input name="fg" type="BSDF" nodename="mix_iridescent_metal_bsdf" />
       <input name="mix" type="float" interfacename="metallic" />
     </mix>