… libraries/pbrlib/genglsl/mx_generalized_schlick_bsdf.glsl
… resources/Materials/Examples/GltfPbr/gltf_pbr_default.mtlx