This PR improves several aspects of the thin film BSDF:
* Adds support for wavelength-dependent complex refraction indices (relevant for metallic base layer)
* Adds (limited) support for thin film effects to `generalized_schlick_bsdf`
* Adds iridescence to glTF PBR

## Wavelength-dependent complex IOR

The current implementation of `thin_film_bsdf` does not support colored base layers, like for example metals. The new implementation supports this case, as shown in the following image.

![colored_base_layer](https://user-images.githubusercontent.com/2821679/185747461-19e6d100-47d1-4190-a6bb-994cac75ebb1.png)

Left: Conductor BSDF with colored reflection.

Center: Thin film on top of the conductor BSDF, old implementation. The yellow tint of the base layer is gone. Only the thin film is visible.

Right: Thin film on top of the conductor BSDF, new implementation. The yellow tint is preserved.

The new implementation now closely follows the Mitsuba reference implementation given in [Belcour and Barla (2017): A Practical Extension to Microfacet Theory for the Modeling of Varying Iridescence](https://belcour.github.io/blog/research/publication/2017/05/01/brdf-thin-film.html). The old implementation was copied from their GLSL shader for BRDF Explorer, which lacked a wavelength-dependent conductor Fresnel term. Note that the authors do not explicitly mention the license of their code. Given the fact that it seemed to be fine to copy their GLSL code, I hope it's fine to copy (parts of) their C++/Mitsuba implementation too.

## Thin film for `generalized_schlick_bsdf`

As mentioned in #1035, the Fresnel term used in `generalized_schlick_bsdf` does not support thin film yet. This changes with this PR.

Given a dielectric base layer, the results between `dielectric_bsdf` and `generalized_schlick_bsdf` are quite close. The following image shows a comparison between thin film over `dielectric_bsdf` (left) and `generalized_schlick_bsdf` (right) BSDFs for a thin film with IOR = 1.33 and thickness = 500 nm over a base layer with IOR = 1.59.

![dielectric_exact_schlick](https://user-images.githubusercontent.com/2821679/185747467-f8362844-bac0-4dc7-9671-2b14af844ccc.png)

However, the implementation has some limitations. It mainly comes from the fact that the Schlick base layer does not consider IOR of the thin film layer. This limitation results in an increase of error if the base layer IOR becomes equal to the thin film IOR. When evaluating the Schlick Fresnel term of the base layer, the incident IOR is assumed to be 1. The following image shows the error by varying the thin film IORs on top of a base layer with IOR = 1.5:

![dielectric_ior_range](https://user-images.githubusercontent.com/2821679/185747471-57a71dcf-2ef9-4fd4-bc8d-a26160358bbe.png)

A similar problem occurs if the base layer is metallic. Here we additionally cannot take the (unknown) complex IOR of the base layer into account for computing the phase shift, thus the error is even worse. Comparison (left: `conductor_bsdf`, right: `generalized_schlick_bsdf`):

![metal_exact_schlick](https://user-images.githubusercontent.com/2821679/185747478-61a5ea32-5434-46b1-8399-09edb3675d9a.png)

Nevertheless, although it has limitations, it conveyes the visual impression of a thin film effect.

## Iridescence for glTF PBR

Recently, the glTF PBR was extended by an iridescence effect ([KHR_materials_iridescence](https://github.com/KhronosGroup/glTF/tree/main/extensions/2.0/Khronos/KHR_materials_iridescence)). Since the MaterialX implementation of glTF PBR is based on `generalized_schlick_bsdf`, it requires the thin film effect for the Schlick Fresnel. Now that this effect is available, it's straightforward to extend the material by thin film. Straightforward except for the [workaround mentioned in #1035](https://github.com/AcademySoftwareFoundation/MaterialX/issues/1035#issuecomment-1196541449), which is required because of limitations in GLSL code generation when referencing the same BSDF multiple times with and without thin-film. Unfortunately, this leads to some duplicated code.