@niklasharrysson 2022-08-24
This looks like a great improvement, thanks @proog128! 

We're planning to take a closer look in the next couple of days. There seems to be some test failures we should investigate.