@jstone-lucasfilm 2022-08-27
@proog128 @niklasharrysson Indeed, this looks like a great improvement, and I'll just highlight one of the issues caught by the test suite.  This looks to be an ESSL restriction that we'd need to take into account in the new code:

```
--- Error log:  D:\a\MaterialX\MaterialX\resources\Materials\Examples\StandardSurfaceTiled_Brass.essl.frag

ERROR: 0:4[82](https://github.com/AcademySoftwareFoundation/MaterialX/runs/7931852759?check_suite_focus=true#step:14:83): '*' :  wrong operand types: no operation '*' exists that takes a left-hand operand of type ' const int' and a right operand of type ' in mediump 3-component vector of float' (or there is no acceptable conversion)
```