@jstone-lucasfilm 2022-09-04
Thanks for putting this proposal together, @proog128, and I've started testing this proposal with existing material examples.  One issue that I'm encountering is that Standard Surface materials seem to lose their thin-film rendering effects with this pull request:

Standard Surface Thin Film (main branch):
![StandardSurfaceThinFilm_Main](https://user-images.githubusercontent.com/6146549/188328913-454e5f89-6346-4316-be51-72a99b6a299d.png)

Standard Surface Thin Film (PR 1055)
![StandardSurfaceThinFilm_PR1055](https://user-images.githubusercontent.com/6146549/188328950-a47acf79-1d37-4e56-aa11-9175752c66b9.png)