@proog128 2022-09-14
@jstone-lucasfilm @niklasharrysson Sorry for the delay. Standard Surface Thin Film is fixed now, it was caused by a typo in the latest refactorings. In addition, I fixed another bug that increased overall brightness of the thin-film Fresnel.