@jstone-lucasfilm 2022-09-17
This looks really promising, thanks @proog128, and I'm now seeing a good visual match with the main branch.

One remaining concern that I have, though, is that the performance for rendering Standard Surface thin-film effects seems to be _reduced_ with respect with the main branch.

To quantify this, I'm attaching screenshots of the standard_surface_thin_film.mtlx material in MaterialXView, with the refresh count set to zero ("--refresh 0"), and with Fraps used to display the frames per second in each build:

Standard Surface Thin Film (main branch):
![MaterialXView_ThinFilm_Main](https://user-images.githubusercontent.com/6146549/190870875-0fe7487e-b9dd-41c5-9eeb-d3cf99f154fc.png)

Standard Surface Thin Film (PR 1055)
![MaterialXView_ThinFilm_PR1055](https://user-images.githubusercontent.com/6146549/190870890-d2a7cc2d-90fa-451e-ab86-93ec7265bf01.png)