@proog128 2022-09-18
The code is quite generic (i.e., it calls the conductor Fresnel function with k = 0 for dielectrics) which is nice but unfortunately problematic for performance. I added a few special cases to improve performance, but it's still not as fast as before. However, given that the BSDF is now more accurate and thus more complicated to compute, I expect it to be slower, at least slightly.

I am measuring the following numbers on my machine:

| | main branch (52b5339) | new impl without optimizations (684e916) | new impl with optimizations (81e8e3b) |
|-|-|-|-|
| Standard Surface Thin Film | 51 fps | 23 fps | 44 fps |
| Thin Film Test 7 (new) | 239 fps | 86 fps | 134 fps |

Thin Film Test 7 is a new test to check the case of `thin_film_bsdf` over `conductor_bsdf`. It is interesting because it cannot take any shortcuts in the implementation.